guacamole: hack: reset display mode on boot
* The display is somehow always set to a wrong mode on boot. The color
is incorrect and the refresh rate is stuck at 60Hz. Normally re-opening
the screen will fix the issue. However, since the display HAL takes the
DRM master role, and since we do not and maybe cannot use OSS display
HAL for this, there is little we can do about this.
* A dirty hack here: by resetting HBM mode and P3 mode, we can force the
kernel DRM driver to reload the correct default mode for the display,
the same effect as turning the display off and back on.